requirements.txt
python-telegram-bot==20.6
python-dotenv==1.0.0
pandas==2.2.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
openpyxl
psycopg2-binary
